Consistent: Detect an april tag and use On-The-Fly path planning to move to it.
Finally, after 3 hours, I found the solution. First, I discovered that if you use mathematical vectors to represent the path with WPILIB's translation2d object, I can easily calculate relative component vectors into absolute component vectors. Previously, I used simple trigonometry with the robots heading as angle to convert components, but this ran into an issue where trig ratios change signs (+/-) in different quadrants, and that the heading goes form +-180 instead of 0 to 360. Lots of sign changing is introduced, and this factor sometimes causes the robot to move in outright wrong paths. However, more time was spent on actually trying to find out why my code (which I didn't touch at all) just wouldn't work, and I hypothesized that it had something to do with photonvisions dashboard settings - and I was right. We discovered the following: in order to use Transform3d to calculate X and Y, you must make sure the processing resolution is high enough (1080x1920) and you must be on the correct processing tab. That means to use the 3d tab, not the 2d tab. It seems that changing the settings on the dashboard actually change how the pi is processing data, so if you are on 2d, you cannot receive any 3d data. Occasionally when you try to switch pipelines or types of things to detect and they don't switch, just reset photon vision. Also the pi gets hot, but it doesn't seem to slow much down??? can't say for sure, did have horrible buffering with dashboard when i reset though.