Updated LED Visual Feedback, Intake Can Now Eject In Intake Position
Before, there was an "intaking" LED colour to signal the driver when the operator was intaking, but this command was overriding the LED colour when it recieves a note, namely being Green, which is the more important visual feedback. I've removed the intaking colour to ensure that in any position, if a note has entered the intake, it will turn green. Code wise, very accurate. Another thing is before, in the intake position, both X and RightBumper sucked in. Updated so that RIghtBumper always ejects, no matter position.