TESTED: Objected Oriented Approach for Note and TpuSystem Class - Working!
The object oriented approach for Notes is now working! The TpuSystem adds/removes/updates as intended when a detection is seen through camera. Using the compareScore method, the best note is supplied. Currently, the area weight factor is the more impactful one - confidence level is a very undependable factor because the AI model itself isn't very great at discerning differences in notes. For example, all orange or orange-red coloured objects are considered notes despite their size, and ringed items that aren't even orange are also considered notes. These fake-notes have high confidence levels, high enough to let you know it needs more training. However, the model is definitely smart enough to discern between a robot bumper and a note. There might be some things that may catch its attention, like perhaps an orange rubber sushi wheel or a thick power cable wrapped in a loop... So yeah, as of now the size has a massive impact on the so-called "score" of a note - this score determines whether the note in question is the best option to go for. These factors need some testing to see if they really work in a competition game or not. Without proper testing, this is probably the best we can get.