Tested: AUTOMATIC NOTE INTAKE COMPLETE! New Upgraded Automatic Intake Commands, Usage of OTF Path
Two new commands: SensorAutoIntake as the name implies drops and spins the intake - the command is finished when the light sensor detects a note; AutoNotePickUp is a Sequential Command which runs SensorAutoIntake first, then retrieves and stops the intake.

AutoNotePickUp is turned into a NamedCommand for PathPlanner to call. Using Sean's upgraded path with event markers, AutoNotePickUp is called at the very beginning, and at the very end (in case no note is actually picked up) the intake is retrieved and stopped.
Additionally, using Rotation Targets allows the robot to face towards the direction of the note when approaching. This ensures high consistency when tracking notes unfavourably to the sides of the bot.
The updated OTF path takes into consideration what happens if no note is detected - when there are no notes or the bestNote is out of range, the robot will only move 10cm forward when the auto is enabled. This is to prevent unexpected and dangerous movement of the robot.

Lastly, added another OTF path method to prepare for April Tag implementation, and minor bug fixes.