Updating Odometry Using Tags, Timed Shooter Sequence
Shooter Sequence is now available. The ShooterSequence Command is called using a trigger, reading a boolean value. When the path nears the end, it changes the value of the boolean and causes the command to run. The ShooterSequence Command cannot be run as an event in the path because the end of the path completely terminates all commands - even when the shooter has just started revving. Additionally, the sequence command could not be run as sequential because the end of the path command is not the end of the robot's motion. Currently, the sequence is programmed to run at 70% of the path's completion. The shooting sequence, ignoring the path, runs for just over 1 second.

AprilTags can now update the robot's odometry. There is an issue where the estimated coordinates in meters do not physically match real world meters; this might be because swapping to Neos changed the base's gear ration and other constant properties required for prediction, but at the same time that should only affect odometry estimation, not value retrieved from april tags. If the coordinates estimated from april tag is not perfectly accurate (april tag coordinates and robot odometry are usually are 10cm LESS for every 1m travelled), then that would mean the given dimensions and angles for the robot's transformation3d relative to the camera (check AprilTagSystem class method) may be incorrect. THIS REQUIRES MORE TESTING, THERE IS NOT ENOUGH INFORMATION TO CONCLUDE THE IMPACT ON ODOMETRY ACCURACY AS OF YET.
Aside from that, the general coordinates and heading, especially heading, are pretty accurately calculated (heading is in sync with robot). April tags are only considered when they are within 2 meters of the robot for less chances of invalid data being received. As of currently, without enough testing and space, it is not quite appropriate to run custom positions yet.

No changes have been made to the Map class as of yet.

Made some minor changes to some classes.