NEW DISCOVERY -> Check details, Small Changes
Moved speaker shooting position buttons to Dpad instead of ABXY. Top is center, and right and left are for source-side and amp-side shooting positions - this depends on the your alliance colour. If you are on the red alliance, your rightside button is amp side, where as on the blue alliance your rightside button is the source. This is to match button direction to the perspective of the driver.

Slightly changed some coordinates for custome positions on map.

Updated AbsoluteDrive so that when using the right and left bumpers for driver controller (for slow turning), also calibrated to absolute coordinates.

Hardware wise, vacuumed the Nav X, reworked usb C power cables for pi (because raspberry pi keeps on getting voltage drops), moved orange pi power cable to the VRM which contains the ethernet switch power so that raspberry pi can have one VRM to itself (to reduce voltage drop by another device)

NEW DISCOVERY: Previously, I thought path planner only returned a destination coordinate to the robot, and the robot would try to go to that coordinate, which was relative to its starting position. But when you think about it, the robot only knows how far and how to get to that destination via its odometry, and if we are updating its  odometry by calibrating it through april tag, then the distance to the final destination is also updated while the path motion is in session. In other words, the robot will realize that the planned destination is a little farther than it initially calculated, so then it speeds up to reach that updated destination within the time limit that it set for itself to reach there. Again, to reword that, lets say the robot thought it traveled 90% of the path already, but then realizes that the final destination is another meter ahead, effectively making its current position only instead of 90. So the robot will decide to accelerate such that it will get to 100% from 80% in the same time it was supposed to get from 90% to 100% in the initially planned path.
This means that the robot auto calibrates its final destination in a path if april tags are available, thus reducing the need for two-step paths. Every motion can be one-step, from A to B and not A to C then C to B.
The acceleration of the robot can sometimes add extra velocity to notes that it shoots out, or mess up timing of commands. This obviously needs further testing.