Not Tested: Path Override, LED Signal For Detection,  Second Try PickUp When Necessary
When joysticks are moved while a path is in session, a new path is created (that barely moves) to override the current.

When a note is detected, the LED on the robot will turn yellow. Drivers can use this as a signal to know whether the TPU is working or not.

Driver station also now prints a boolean representing whether notes are being detected and the photon camera is on.

NotePickUp command now takes in a boolean to determine whether the note pick up auto should try again if a note was missed.

note pick up auto now has a backward bezier tangent to prevent the robot from pushing away the note before it picks it up.