Note Detected LED Indication and Path Overriding, Unresolved Note PickUp Scenarios
When a valid note is detected (in bound, or has a score > 0, or the intake is free), the LEDs will turn yellow. For discrepancy, idle colour is now blue instead of rainbow. Intaking a note still shows green. This can be updated such that the closer the robot is to the note, the warmer the colour of the LED, though this is not a high priority change.

Drivers can now input any joystick direction to stop and override paths in case of emergency. The override works by scheduling another path command while the robot is still completing the original path. The newly scheduled path is very short, and quickly returns control back to the joysticks upon completion. Overriding will return the intake to the rest position automatically.

New custom poses were added to map, though not tested for accuracy.

Camera transformation FROM robot slightly altered to accommodate for 2d printed casing.

For some reason, odometry was off again and I found a new number for gear ratio.

UNRESOLVED: Before, there were two scenarios where the note auto could not consistently pick up notes - right in front of robot, and extremely to the sides outside camera FOV.
Normally, if the intake is already down and moving towards a note, the note will be pushed up into the intake by the bottom set of wheels. However, when the note is directly in front of the bumper, the intake lands on top of the note and is pushed away rather than consumed. To resolve this, I initially had the robot retreat. In a 3-pose path, the robot would drop the intake when retreating, and scoop the intake when approaching, thus resolving the problem. However, when I was testing, this path had a lot of errors (likely because of the logic in the path I was using) where it became very inconsistent to pick up notes on the sides. When removing the retreat, the path had no issue picking up side notes, but the retreat path could not. There are simple fixes to this issue, such as more testing and removing bezier curves, or turning it into two step paths - but these will require further effort. As for now, the mitigation is simply using a slower acceleration and velocity, praying notes are picked up.
For notes on the very side which exit the camera FOV, I initially thought a "miss" would push the note into the view of the camera again, but this is simply not always true. Attempting to resolve this produced two problems- one, a "searching" command would need to be in place after a miss. For example, it would turn 360 until it found a note, or it should retreat until notes enter its FOV. The second problem is actually figuring out how to run different paths after each other like commands. The problem with paths is that their command calls are instant, but the motion of the robot finishing the path ends way later. Paths do not tell you when they are complete, or interrupted, so we would need to modify boolean variables as event markers to notify following commands. And, command scheduling or creating orders is not always as simple as it appears... Indeed, with enough time to test and debug this would be no problem, but it was not considered to spend lots of time "perfecting" or engineering a product in the schedule because the end goal was simply set to  "get this usable for CNE" and not "make it consistent for CNE".

This is just after thought, but end goals determine how you lay out schedules and tasks. With such a vague goal and for CNE, as well as being considerate for time, I made a schedule which only got the basic features working, usable indeed but not always consistent. In terms of "engineering" and constant testing or debugging, the schedule I made only moderately considered time for that. However, errors from early on in design not resolved via testing are bound to affect future progress someway or another. Doing sufficient testing, ESPECIALLY WITH RECORDED DATA, is one of the ways to analyze and thus resolve problems to the best of our ability, ensuring the best foundation for the next task. Half-assing tasks now will half-ass everything, and definitely depends on just how "engineered" you want to make the project. Lastly, with proper managing and planning skills, SMART goals can be achieved smoothly - that includes making sure team members are working to the best of their ability. Being a manager is much harder than being a worker because you are managing the work of more than just yourself. If you don't put a ton of effort into leading a project, then you can't expect a ton effort from team members in return. Running into problems, falling behind schedule, things not going smoothly, and times of wanting to give up sometimes bring me debates, and often a reflection on things I could have done better or changed. In tight times like these, its important to view your actions from a third person perspective, with logic, and no bias of emotion. A long day at the club can get you fed up, but always remember to keep your cool if you truly want to progress a project to the best of your ability.